+++ docs/changelog.md
@@ -13,6 +13,10 @@ As features stabilize some brief notes about them will accumulate here.
 
 * Not yet
 
+#### New
+
+* Color schemes: [DanQing Light (base16)](colorschemes/d/index.md#danqing-light-base16), [SynthwaveAlpha](colorschemes/s/index.md#synthwavealpha), [SynthwaveAlpha (Gogh)](colorschemes/s/index.md#synthwavealpha-gogh)
+
 ### 20220807-113146-c2fee766
 
 #### New